+++ packages/lucid/test/specs/hello.ts
@@ -1,10 +1,10 @@
 import { Console, Effect, Logger, LogLevel, pipe, Schedule } from "effect";
-import { HelloContract, User } from "./services";
+import { HelloContract, User } from "./services.js";
 import { Constr, Data } from "@lucid-evolution/plutus";
 import { getAddressDetails } from "@lucid-evolution/utils";
 import { fromText } from "@lucid-evolution/core-utils";
-import { TxSignBuilder } from "../../src";
-import { handleSignSubmit, withLogRetry } from "./utils";
+import { TxSignBuilder } from "../../src/index.js";
+import { handleSignSubmit, withLogRetry } from "./utils.js";
 
 const DatumSchema = Data.Object({
   owner: Data.Bytes(),