+++ packages/lucid/test/specs/services.ts
@@ -2,10 +2,11 @@ import { Config, Context, Effect, Layer, pipe } from "effect";
 import {
   applyDoubleCborEncoding,
   Blockfrost,
+  Koios,
   Lucid,
   SpendingValidator,
   validatorToAddress,
-} from "../../src";
+} from "../../src/index.js";
 import scripts from "./contracts/plutus.json";
 
 const makeUser = Effect.gen(function* ($) {
@@ -14,9 +15,11 @@ const makeUser = Effect.gen(function* ($) {
     Config.string("VITE_BLOCKFROST_KEY"),
     Config.string("VITE_SEED"),
   ]);
-  const user = yield* Effect.tryPromise(() =>
-    // Lucid(new Kupmios("http://localhost:1442", "ws://localhost:1337"), "Preview")
-    Lucid(new Blockfrost(apiURL, apiKey), "Preprod"),
+  const user = yield* Effect.tryPromise(
+    () =>
+      // Lucid(new Kupmios("http://localhost:1442", "ws://localhost:1337"), "Preview")
+      Lucid(new Blockfrost(apiURL, apiKey), "Preprod"),
+    // Lucid(new Koios("https://preprod.koios.rest/api/v1"),"Preprod")
   );
   user.selectWallet.fromSeed(seed);
   return {